Content-type for fallback-data ved manglende lokalisert innhold. Tar i bruk denne for skjemaoversikter
