+++ src/main/resources/site/content-types/localized-content-data-fallback/localized-content-data-fallback.ts
@@ -1,0 +1,37 @@
+// WARNING: This file was automatically generated by "no.item.xp.codegen". You may lose your changes if you edit it.
+export interface LocalizedContentDataFallback {
+  /**
+   * Velg innholdstyper
+   */
+  contentTypes: Array<string>;
+
+  /**
+   * Verdier
+   */
+  items?: Array<{
+    /**
+     * Velg innhold
+     */
+    contentId: string;
+
+    /**
+     * Aktivert
+     */
+    enabled: boolean;
+
+    /**
+     * Tittel
+     */
+    title: string;
+
+    /**
+     * Ingress
+     */
+    ingress: string;
+
+    /**
+     * Overstyr URL
+     */
+    targetUrl?: string;
+  }>;
+}
