+++ src/main/resources/types/content-types/content-config.d.ts
@@ -51,6 +51,7 @@ import { FormsOverview } from '../../site/content-types/forms-overview/forms-ove
 import { FragmentCreator } from '../../site/content-types/fragment-creator/fragment-creator';
 import { PageMeta } from 'site/content-types/page-meta/page-meta';
 import { SearchConfigV2 } from '../../site/content-types/search-config-v2/search-config-v2';
+import { LocalizedContentDataFallback } from '../../site/content-types/localized-content-data-fallback/localized-content-data-fallback';
 
 type CustomContentDataConfigsWithoutDescriptor = {
     'animated-icons': AnimatedIcons;
@@ -69,6 +70,7 @@ type CustomContentDataConfigsWithoutDescriptor = {
     'guide-page': GuidePage;
     'internal-link': InternalLink;
     'large-table': LargeTable;
+    'localized-content-data-fallback': LocalizedContentDataFallback;
     'main-article': MainArticle;
     'main-article-chapter': MainArticleChapter;
     'megamenu-item': MegamenuItem;
